Tracks Auto-Sizing
The tracks now auto-size themselves based on the camera's distance from
it's target body. We should change this later to be based on it's
distance from each object (so that we can also do turning off-and-on
based on distance).

Also increased the clip size to 100000 and the zoom max to 9500.