Organizing Resources for Building
Created a resources folder and copied all of the relevant text files to
it. The code is not yet updated to handle this, but Jihad is going to
work on it. Once the code is updated, the build issues with missing text
files should be solved, as the Resources folder goes into the Build Data
folder.