Distant Planet Texture/Material/Prefab Added
I added a new material, texture, and prefab for indicating planets
and/or moons at a distance. This will be used in the contextual-distance
rendering scripts to lower required processing power, and make objects
clickable at a distance. Currently they are not yet attached to a
script, but prepared to be.