Distant Planet Icons - Clickable, Self-Scaling, Context Rendering
I've added icons to each planet (not the moons or the sun yet). These
icons are not rendered while the camera is close to an object, but if it
is far away (at about the point that the planet becomes too small to
click on) the icon appears and stays at a good size at any distance.

These icons can be double clicked like the planets, and will make the
camera switch targets to the icon's planet just like double clicking on
the planet itself.

Known Bugs:
The icons are not clickable beyond a certain distance (very far). Not
sure why this is yet - I think it has to do with the length of the
raycast.

I also changed the CameraUserControl script to shoot a raycast each
update (instead of just when clicking) which will allow us to do
mouse-over events in addition to clicking events.