Added Materials to HSS and Colliders to All Models
I added textures and scripts to the Hubble telescope prefab, and I also
added colliders to all of the models such that they can be clicked on
and targeted.

Additionally, I think I've found the cause of the blue-dot near the sun
bug. The standard-moon object is used to create all of the moons when
the system starts up, but the standard model stays in the space and
isn't destroyed after being used. We should be using a prefab for this
instead of an object in the map (this would solve the bug).