Bug Fix: Ship Texture Glitch
Bug: Most of the new ships would show up with black textures when
instantiated in Voyager at run-time (with the Create Ship GUI).
Fix: Apparently this issue related to the scale of each of the ships. If
they are large enough, the textures render without issue. I changed the
scale size of each one in the script to be the smallest size possible
for rendering to occur properly. This does not, however, solve the long
term problem of making sure the ships are scaled as in real life, and
still textured.
Also - I added some parent objects to the ships, to make sure their
center of mass is where the camera focuses on them.