Bug Fix: Info Window Format & Bad Text Input
Bug: Right clicking on a planet or object would sometimes result in a
'bad text input' error in the GUI function to create an infoWindow. This
would prevent further right clicking to create windows and throw an
error every update. Additionally, the 'x' button was randomly placed in
the window, and the window could be dragged by clicking and dragging
anywhere.
Fix: The atmosphere's of the planets and objects were not on an
'ignore-raycast' layer (I had previously thought they were). I put them
all on this layer to ensure that right clicking didn't accidentally
select one of these atmospheres instead of the object themselves.
Additionally, I moved the 'x' box into the top right corner of the
windows and rescaled it to fit. I also made the windows draggable only
by the top bar, which is more fitting with the feel we're going for.
This issue has been resolved.