Info Window Update & Prep For Standard Camera Angles
This update adds several buttons to the right-click context windows in
preparation for the standard camera angle functionality. It also makes
only the top bar drag-able.
I've also commented out a debug note from shipOEHistory.
Finally, I added a function which is meant to take the desired standard
camera angle, and change the position of the camera. Right now this
function is not working, and not being called from anywhere.