Prep work for Ilya
The maneuver node button and window has been replaced with simply a mean
anomaly entry (in degrees). The code to convert to radians is in the
file. Currently, inputting this value does nothing, but Ilya will be
making use of the value once he has set up the Mean Anomaly burn list.