Visualizing parking orbits. Fixed flight-angle.
The visualization of parking orbits has been fixed for the most part.
Some slight issues might still be there.
The equation for the flight-angle had the wrong variable, this has been
fixed.
The Time window can be dragged. The slider doesn't move with it, but it
can be dragged out of the way if need be.