Fixed true anomaly bug. Started Hyperbolic orbits.
Sometimes the expression for arcCos is a tiny fraction larger than 1,
which causes an error. To Fix it, it's been hard-pulled to 1.
For hyperbolic orbits, put in the calculation for v (true anomaly), E
(eccentric anomaly), M (mean anomaly) and n (mean motion). But the
function findPos causes an error with epsilon crash.